requirements.txt
# --- Core & UI ---
gradio==6.3.0
pydub
soundfile
python-dotenv
psutil

# --- STT (Whisper Turbo) & TTS (Parler) ---
# Alignement sur CUDA 12.4
torch --index-url https://download.pytorch.org/whl/cu124
torchaudio --index-url https://download.pytorch.org/whl/cu124
transformers==4.46.1
accelerate>=1.1.0
sentencepiece
protobuf==4.25.8
# Installation via Git pour avoir les dernières optimisations multilingues
git+https://github.com/huggingface/parler-tts.git

# --- LLM & MCP ---
mcp==1.25.0
uv==0.9.26
httpx
# Pour connect LLM Studio
openai==2.15.0
# Pour connect local (Llama.cpp)
llama-cpp-python==0.3.16
huggingface_hub==0.36.0

# --- Audio Utils & Fallbacks ---
descript-audiotools==0.7.4
kokoro==0.9.4
onnxruntime-gpu==1.20.1
librosa